更新

+++ BE/src/main/java/com/example/calmall/cartitem/entity/CartItem.java
@@ -3,20 +3,26 @@ package com.example.calmall.cartitem.entity;
 import jakarta.persistence.*;
 import lombok.Data;
 
-// ユーザーのカート情報を管理する中間エンティティ
+/**
+ * ユーザーのカート情報を管理する中間エンティティ
+ */
 @Entity
 @Data
 public class CartItem {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id; // カートアイテムID
+    private Long id;
 
-    private Integer userId; // 所属ユーザーID
+    // 所属ユーザーID（User.id を参照）
+    private Long userId;
 
-    private String itemCode; // 商品ID（楽天形式）
+    // 商品ID（楽天形式）
+    private String itemCode;
 
-    private Integer quantity; // 数量0になったら削除
+    // 数量（0になったら削除）
+    private Integer quantity;
 
-    private String option; // APIから与えられる商品オプション
+    // APIから与えられる商品オプション
+    private String option;
 }
